session/llama2/requirements.txt
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.41.1
transformers==4.33.1
git+https://github.com/malaysia-ai/trl
tokenizers==0.13.3
deepspeed==0.10.0
lightning==2.0.7
lightning-cloud==0.5.37
lightning-utilities==0.9.0
pytorch-lightning==2.0.7
wandb==0.15.8
evaluate==0.4.0
scipy==1.11.2